pyproject.toml
[project]
name = "arti-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.27.0",
    "hatchling>=1.29.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "textual>=8.1.1",
    "docker>=7.0.0",
]

[project.scripts]
arti-ops = "arti_ops.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_file = "logs/pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
